# Conflicts:
#	Service/CustomerService.java

+++ DAO/AdminDAO.java
@@ -57,7 +57,7 @@ public class AdminDAO implements GenericDAO<Admin> {
         return admin.getPassword();
     }
 
-    public void createNewAdmin(String username, String password, Date dateOfBirth,String role,int workingHours){
+    public void createNewAdmin(String username, String password, Date dateOfBirth,String role,double workingHours){
         Admin admin=new Admin(username,password,dateOfBirth,role,workingHours);
         Database.admins.add(admin);
