+++ Entity/Admin.java
@@ -26,7 +26,7 @@ public class Admin extends User{
         this.workingHours=workingHours;
     }
 
-    public Admin(String username, String password, Date dateOfBirth, String role, int workingHours) {
+    public Admin(String username, String password, Date dateOfBirth, String role, double workingHours) {
         super(username,password,dateOfBirth);
         this.role=role;
         this.workingHours=workingHours;
